Add objects!
This is very messy, and the classes all seem to end up in whatever
package the user is in once they're generated, but they seem to be
working.  Note: This commit doesn't include parshing methods under
the classes other than constructors.  Hopefully that will mostly
be a matter of hooking code up to other code.